add test project. refactor playback/record into RecordingHandler to ease testing
