+++ .gitignore
@@ -457,6 +457,9 @@ restoredPackages/
 src/dotnet/Mgmt.CI.BuildTools/NugetToolsPackage/CI.Tools.Package/build/tasks/net461/
 src/dotnet/Mgmt.CI.BuildTools/NugetToolsPackage/CI.Tools.Package/build/tasks/netstandard2.0
 
+# ensure local debugging of the test-proxy will not end up committed
+tools/test-proxy/Azure.Sdk.Tools.TestProxy/recordings/
+
 # local settings
 .DS_Store
 .vscode
