+++ tools/test-proxy/Azure.Sdk.Tools.TestProxy.Tests/AdminFunctionalityTests.cs
@@ -1,0 +1,14 @@
+using System;
+using Xunit;
+
+namespace Azure.Sdk.Tools.TestProxy.Tests
+{
+    public class AdminFunctionalityTests
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
