+++ tools/test-proxy/Azure.Sdk.Tools.TestProxy.Tests/RecordingHandlerTests.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Xunit;
+
+namespace Azure.Sdk.Tools.TestProxy.Tests
+{
+    public class RecordingHandlerTests
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
