+++ tools/test-proxy/Azure.Sdk.Tools.TestProxy.Tests/TransformTests.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using Xunit;
+
+namespace Azure.Sdk.Tools.TestProxy.Tests
+{
+    public class TransformTests
+    {
+
+
+        [Fact]
+        public void CheckMSALMatcher()
+        {
+            
+        }
+    }
+}
